+++ lib/src/models/epub_save_options_data.dart
@@ -382,6 +382,12 @@ class EpubSaveOptionsData extends HtmlSaveOptionsData {
       prettyFormat = null;
     }
 
+    if (json.containsKey('RemoveJavaScriptFromLinks')) {
+      removeJavaScriptFromLinks = json['RemoveJavaScriptFromLinks'] as bool;
+    } else {
+      removeJavaScriptFromLinks = null;
+    }
+
     if (json.containsKey('ReplaceBackslashWithYenSign')) {
       replaceBackslashWithYenSign = json['ReplaceBackslashWithYenSign'] as bool;
     } else {